requirements.txt
psutil
requests
GPUtil
psycopg2-binary
